requirements.txt
arabic-reshaper==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
decorator==4.4.2
gTTS==2.5.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==1.0.3
numpy==2.2.4
pillow==10.4.0
proglog==0.1.11
python-bidi==0.6.6
python-dotenv==1.1.0
requests==2.32.3
tqdm==4.67.1
urllib3==2.3.0
